first pass, no fixes
