+++ .eslintrc.json
@@ -17,7 +17,8 @@
     "react-hooks",
     "prettier",
     "jsx-a11y",
-    "react-hook-form"
+    "react-hook-form",
+    "header"
   ],
   "settings": {
     "react": {
@@ -59,6 +60,29 @@
     {
       "files": ["*.e2e.ts"],
       "extends": ["plugin:playwright/playwright-test"]
+    },
+    {
+      "files": ["*.{ts,tsx}"],
+      "excludedFiles": [
+        "*.spec.ts",
+        "*.e2e.ts",
+        "*.stories.tsx",
+        "tools/**/*",
+        "*.config.ts"
+      ],
+      "rules": {
+        "header/header": [
+          2,
+          "line",
+          [
+            " This Source Code Form is subject to the terms of the Mozilla Public",
+            " License, v. 2.0. If a copy of the MPL was not distributed with this",
+            " file, you can obtain one at https://mozilla.org/MPL/2.0/.",
+            "",
+            " Copyright 2023 Oxide Computer Company"
+          ]
+        ]
+      }
     }
   ]
 }
