+++ package.json
@@ -108,6 +108,7 @@
     "esbuild-register": "^3.4.2",
     "eslint": "^8.42.0",
     "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-playwright": "^0.12.0",
     "eslint-plugin-prettier": "^4.0.0",
